+++ src/service/color.js
@@ -9,9 +9,9 @@ export function rgbToHex( r, g, b ) {
 
 export function hexToRgb( hex ) {
   var result = /^#?([a-f\d]{2})([a-f\d]{2})([a-f\d]{2})$/i.exec( hex );
-  return result ? {
-    red: parseInt( result[ 1 ], 16 ),
-    green: parseInt( result[ 2 ], 16 ),
-    blue: parseInt( result[ 3 ], 16 )
-  } : null;
+  return result ? [
+    parseInt( result[ 1 ], 16 ),
+    parseInt( result[ 2 ], 16 ),
+    parseInt( result[ 3 ], 16 )
+  ] : null;
 }